pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dcheck-enterprise-runner"
version = "0.1.0"
description = "Enterprise Runner for DCheck: YAML-driven batch execution with JSON outputs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
dcheck-enterprise-runner = "dcheck_enterprise_runner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
